refactor: 将NUM_CONCAT类型从int改为int64_t
修改NUM_CONCAT常量类型以保持与DIM_LAST类型一致，避免潜在的整数溢出风险

+++ examples/inductor_pattern_match/torch_op_plugin/sub_mul_concat.cpp
@@ -20,7 +20,7 @@ See the License for the specific language governing permissions and
 #include "torch_npu_helper.h"
 
 namespace acl_ops {
-static constexpr int NUM_CONCAT = 4;
+static constexpr int64_t NUM_CONCAT = 4;
 static constexpr int64_t DIM_LAST = 2;
 
 /// This operator will concat four tensors: lhs, rhs, lhs - rhs, lhs * rhs, along the last dim.
